+++ src/components/LibraryList.js
@@ -9,36 +9,49 @@ import { Link } from 'react-router-dom';
 const LibraryList = () => {
   const dispatch = useDispatch();
   const libraries = useSelector(state => state.libraries);
-  const [editingLibrary, setEditingLibrary] = useState({ id: '', name: '', ownername: '', books: [] });
+  const [editingLibrary, setEditingLibrary] = useState({ id: '', name: '', ownerName: '', books: [] });
   const [selectedLibrary, setSelectedLibrary] = useState(null);
 
   useEffect(() => {
     axios.get('http://localhost:8080/library')
-    .then(res => {
-      console.log("data comes here", res.data);
-      dispatch(setLibraries(res.data));
-    })
-    .catch(error => console.log(error));
-   
-  }, [dispatch]);
-  const handleDelete = (id) => {
-    axios.delete(`http://localhost:8080/library/${id}`)
       .then(res => {
-        console.log("hiiiiiii");
-        dispatch(deleteLibrary(id));
-     
+        dispatch(setLibraries(res.data));
       })
       .catch(error => console.log(error));
+  }, []);
+
+  const handleEditLibrary = (selectedLibrary) => {
+    setEditingLibrary(selectedLibrary);
   };
 
-  const handleUpdateClick = (library) => {
-    setSelectedLibrary(library);
-    setEditingLibrary(library);
+  const handleUpdateLibrary = (id) => {
+    const confirmUpdate = window.confirm('Are you sure you want to update this library?');
+    if (confirmUpdate) {
+      if (editingLibrary.name.trim() === '' || editingLibrary.ownerName.trim() === '') {
+        alert('Please fill the data correctly');
+        return;
+      }
+      axios.put(`http://localhost:8080/library/${id}`, editingLibrary)
+        .then(res => {
+          const updatedLibrary = { ...editingLibrary };
+          dispatch(updateLibrary(updatedLibrary));
+          setEditingLibrary({ id: '', name: '', ownername: '', books: [] });
+          alert('Library updated successfully');
+        })
+        .catch(error => console.log(error));
+    }
   };
 
-  const handleCancelUpdate = () => {
-    setSelectedLibrary(null);
-    setEditingLibrary({ id: '', name: '', ownername: '', books: [] });
+  const handleDeleteLibrary = (id) => {
+    const confirmDelete = window.confirm('Are you sure you want to delete this library?');
+    if (confirmDelete) {
+      axios.delete(`http://localhost:8080/library/${id}`)
+        .then(res => {
+          dispatch(deleteLibrary(id));
+          alert('Library deleted successfully');
+        })
+        .catch(error => console.log(error));
+    }
   };
 
   return (
@@ -55,26 +68,29 @@ const LibraryList = () => {
         </thead>
         <tbody>
           {libraries && libraries.map(library => (
-            <tr key={library.id}>
-              <td>{library.id}</td>
-              <td>
-               
-                  {library.name}
-              
-              </td>
-              <td>{library.ownerName}</td>
-              <td>
-               
-              <button>
-              <Link to={`/books/${library.id}`}>Show Books</Link>
-             
-  </button>
-                
-                  <button onClick={() => handleUpdateClick(library)}>Update</button>
-                
-                <button onClick={() => handleDelete(library.id)}>Delete</button>
-              </td>
-            </tr>
+            editingLibrary.id === library.id ? (
+              <tr key={library.id}>
+                <td>{library.id}</td>
+                <td><input type="text" value={editingLibrary.name} onChange={(event) => setEditingLibrary({ ...editingLibrary, name: event.target.value })} /></td>
+                <td><input type="text" value={editingLibrary.ownerName} onChange={(event) => setEditingLibrary({ ...editingLibrary, ownerName: event.target.value })} /></td>
+                <td>
+                  <button onClick={() => handleUpdateLibrary(library.id)} className="btn btn-success">Update</button>
+                  <button onClick={() => setEditingLibrary({ id: '', name: '', ownerName: '', books: [] })} className="btn btn-warning">Back</button>
+                </td>
+              </tr>
+            ) : (
+              <tr key={library.id}>
+                <td>{library.id}</td>
+                <td>{library.name}</td>
+                <td>{library.ownerName}</td>
+                <td>
+                <button><Link to={`/addBook/${library.id}`}  className="btn btn-sm btn-info" >Add Book</Link></button>
+                  <button><Link to={`/books/${library.id}`}>Show Books</Link></button>
+                  <button onClick={() => handleEditLibrary(library)}>Edit</button>
+                  <button onClick={() => handleDeleteLibrary(library.id)}>Delete</button>
+                </td>
+              </tr>
+            )
           ))}
         </tbody>
       </table>
@@ -82,7 +98,7 @@ const LibraryList = () => {
       {selectedLibrary && (
         <LibraryUpdateForm
           library={selectedLibrary}
-          onClose={handleCancelUpdate}
+          onClose={() => setSelectedLibrary(null)}
         />
       )}
     </div>
